adding lpcxpresso55s69
